Implement root Flask blueprint
http://flask.pocoo.org/docs/0.10/blueprints/
